+++ FMIndex/SSA.h
@@ -55,9 +55,9 @@ public:
   uint length();
 
   uint LF(uint i);
-  uint locate_id(uchar *pattern, uint m);
-  uint locate(uchar *pattern, uint m, size_t **occs);
-  uint locateP(uchar *pattern, uint m, size_t *left, size_t *right,
+  unsigned long locate_id(uchar *pattern, uint m);
+  unsigned long locate(uchar *pattern, uint m, size_t **occs);
+  unsigned long locateP(uchar *pattern, uint m, size_t *left, size_t *right,
                size_t last);
 
   uchar *extract_id(uint id, uint *strLen, uint max_len);
